Converts bitboard to bigint

+++ db/migrate/20140815153710_create_boards.rb
@@ -2,7 +2,7 @@ class CreateBoards < ActiveRecord::Migration
   def change
     create_table :boards do |t|
       t.integer :user_id
-      t.integer :bitboard
+      t.bigint :bitboard
 
       t.timestamps
     end
